src/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers==4.6.0
djangorestframework
idna
pytz
pyYAML
requests
sqlparse
urllib3